User entity created to handle registration and setting up endpoint to ask for an email activation using a uuid as a token

+++ server/src/main/java/com/codecool/sv_server/dto/SignupResponseDto.java
@@ -1,4 +1,4 @@
 package com.codecool.sv_server.dto;
 
-public record SignupResponseDto(String email, int id){
+public record SignupResponseDto(String email, long id){
 }
